// src/models/customer.ts
export interface Customer {
  customer_id: string;
  first_name: string;
  last_name: string;
  password_entered: string;
  password: string;
  email: string;
  mobilenumber: string;
  resgistration_type: string;
  dob: string;
  status: string;
  otp: string;
  otp_status: string;
  mail_encode: string;
  mail_verify: string;
  created: string;
  modified_date: string;
  appdevice_id: string;
}